requirements.txt
openaq==0.5.0
pandas>=2.3.1
matplotlib>=3.10.5
python-dotenv>=1.0.0
flask>=3.1.2
scikit-learn>=1.7.1
jupyter>=1.1.1
xgboost>=3.1.2
plotly>=6.3.0